Update index.md
Ah yes, this is one of the labs that I tried to have ChatGPT fix the formatting of the lab and turn it into markdown and it changed some of the code.  I had this fixed in my working repo.  ChatGPT made it look great but it didn't work anymore.